refactor: added section pane
