+++ apps/antalmanac/src/components/RightPane/AddedCoursePane/CustomEventDetailView.tsx
